scripts/upload/pyproject.toml
[project]
name = "upload"
version = "0.1.0"
description = "Upload LoopNet listings to Supabase"
requires-python = ">=3.12"
dependencies = [
    "supabase>=2.10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "geopy>=2.4.0",
]
